Test-confirmed version
This version of the code changes how the movement structures for
vertically moving the wrist and the elbow work, sprinkling delays
around. Confirmed in testing to work on the arm with this version of
code, arm moved up and down as told. Presumably base as well, but it
wasn't actually tested as the dynamixel was having...issues